requirements.txt
mysql-connector-python==9.1.0
python-dotenv==1.0.1
pandas==2.3.2
openpyxl==3.1.5